+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Add `show_studio` to search index for seasons and episodes
 
 ## [0.8.5-beta] - 2024-01-30
 ### Added